requirements.txt
apispec==5.1.1
apispec-webframeworks==0.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cobble==0.1.3
colorama==0.4.4
config==0.5.1
docx2pdf==0.1.8
docx2txt==0.8
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
lxml==4.8.0
mammoth==1.4.19
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.8
python-docx==0.8.11
PyYAML==6.0
requests==2.27.1
SQLAlchemy==1.4.32
tqdm==4.64.0
urllib3==1.26.9
uuid==1.30
Werkzeug==2.0.3
